replace delegation of RSocketState with direct access
